refactor: Update dependencies and environment variables
This commit updates the dependencies in the package.json file to their latest stable versions. Specifically, it updates the following dependencies:
- mongoose from version 8.4.5 to 8.5.3
- winston from version 3.13.0 to 3.14.2

Additionally, it modifies the .env.sample file to remove the BCRYPT_SALT_ROUNDS variable and adds the following optional environment variables:
- REFRESH_TOKEN_EXPIRY with a value of 7d

These changes ensure that the project is using the latest dependencies and includes the necessary environment variables for token expiration.